Refactor reducers and state (#59)
* Added AppDataReducer

* Added AppDataContext

* Added TextInputHeightContext

* Added CalendarReducer

* Added CategoryReducer

* Added LimitReducer

* Replaced selected limit object with id

* Fixed eslint warnings
