+++ App.tsx
@@ -2,12 +2,13 @@ import "./src/extensions";
 import React from "react";
 import { KeyboardAvoidingView, Platform, View, ViewStyle } from "react-native";
 import * as SystemUI from "expo-system-ui";
-import TextInputHeight from "./src/helpers/TextInputHeight";
+import AppDataContext from "./src/helpers/AppDataContext";
+import AppDataProvider from "./src/helpers/AppDataProvider";
+import TextInputHeightProvider from "./src/helpers/TextInputHeightProvider";
 import HeaderView from "./src/views/HeaderView";
 import AppbarView from "./src/views/AppbarView";
 import CalendarView from "./src/views/CalendarView";
 import CategoryView from "./src/views/CategoryView";
-import AppDataProvider, { AppDataContext } from "./src/helpers/AppDataProvider";
 import LimitView from "./src/views/LimitView";
 import GraphView from "./src/views/GraphView";
 import { COLORS } from "./src/constants";
@@ -29,7 +30,7 @@ export default function App() {
     <AppDataProvider>
       <View style={{ flex: 1 }}>
         <HeaderView />
-        <TextInputHeight.Provider>
+        <TextInputHeightProvider>
           <KeyboardAvoidingView
             style={{ flex: 1 }}
             enabled={Platform.OS === "ios"}
@@ -40,7 +41,7 @@ export default function App() {
             <LimitView style={pageStyle["limits"]} />
             <View style={pageStyle["settings"]} />
           </KeyboardAvoidingView>
-        </TextInputHeight.Provider>
+        </TextInputHeightProvider>
         <AppbarView page={page} setPage={setPage} />
       </View>
       <AppDataContext.Consumer>
