+++ src/helpers/TextInputHeightContext.ts
@@ -1,0 +1,5 @@
+import React from "react";
+
+const TextInputHeightContext = React.createContext(54);
+
+export default TextInputHeightContext;
